+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [1.0.7] - 09/23/2020
+* Update README
+
+## [1.0.6] - 09/23/2020
+* Added reload, getInstance methods
+* Add option to pass custom SharedPreferences instance
+* Update example project
+
 ## [1.0.5] - 11/24/2019
 * Added clear option
 
@@ -14,9 +22,7 @@
 * Added example
 
 ## [1.0.0] - 11/22/2019
-
 * Added null safety for getString.
 
 ## [0.0.1] - 11/22/2019
-
 * initial release.